added route for tables/new and added  CreateTableForReservations
