+++ front-end/src/layout/Routes.js
@@ -5,6 +5,7 @@ import Dashboard from "../dashboard/Dashboard";
 import NotFound from "./NotFound";
 import { today } from "../utils/date-time";
 import CreateReservation from "../Reservations/CreateReservation";
+import CreateTableForReservation from "../tables/CreateTableForReservation";
 
 /**
  * Defines all the routes for the application.
@@ -27,6 +28,9 @@ function Routes() {
       </Route>
       <Route path="/reservations/new">
         <CreateReservation />
+      </Route>
+      <Route path="/tables/new">
+        <CreateTableForReservation />
       </Route>
       <Route>
         <NotFound />
